+++ portico_webworks_plugin.php
@@ -2,17 +2,18 @@
 /**
  * Plugin Name: Portico Webworks Hotel Website Manager
  * Description: Portico Webworks plugin.
- * Version: 0.3.0
+ * Version: 0.3.1
  * Author: Portico Webworks
  * Author URI: https://porticowebworks.com
+ * License: Proprietary (All Rights Reserved) - Unauthorized use by other companies or for any purposes is prohibited without written permission.
  */
 
 if (!defined('ABSPATH')) {
 	exit;
 }
 
-define('PORTICO_WEBWORKS_PLUGIN_FILE', __FILE__);
-define('PORTICO_WEBWORKS_VERSION', '0.3.0');
+define('PW_PLUGIN_FILE', __FILE__);
+define('PW_VERSION', '0.3.1');
 
 require_once __DIR__ . '/includes/admin-page.php';
 require_once __DIR__ . '/includes/property-post-type.php';